% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{dnormmix}
\alias{dnormmix}
\title{This function computes the PDF of a mixture distribution.}
\usage{
dnormmix(x, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{x}{a numeric value where the PDF is evaluated.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{The probability of the first distribution occurring.}
}
\value{
A numeric value for the PDF at the original `x`.
}
\description{
This function computes the PDF of a mixture distribution.
}
