% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{myqnormmix}
\alias{myqnormmix}
\title{Vectorized quantile function}
\usage{
myqnormmix(p, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{p}{A vector of probabilities used to calculate different quantiles.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{The probability of the first distribution occurring.}
}
\value{
A vector of quantiles that corresponds to their respective probabilities of p
}
\description{
This function is a vectorized version of the qnormmix function
}
