% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{pnormmix}
\alias{pnormmix}
\title{Mixture CDF function with user-defined parameters}
\usage{
pnormmix(q, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{q}{A numeric value where the CDF is evaluated.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{The probability of the first distribution occurring.}
}
\value{
A numeric value of the CDF at the original `q`.
}
\description{
This function computes the CDF of a mixture distribution.
}
