% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{qnormmix}
\alias{qnormmix}
\title{This function generates a random sample from a mixture of two normal distributions.}
\usage{
qnormmix(p, mean1, sd1, mean2, sd2, mixprob)
}
\arguments{
\item{p}{a value for the quantile to be calculated.}

\item{mean1}{Mean of the first normal distribution.}

\item{sd1}{Standard deviation of the first normal distribution.}

\item{mean2}{Mean of the second normal distribution.}

\item{sd2}{Standard deviation of the second normal distribution.}

\item{mixprob}{The probability of the first distribution occurring.}
}
\value{
the corresponding quantile to the input p
}
\description{
This function generates a random sample from a mixture of two normal distributions.
}
